add default language folder, refactor c4 and gopher quality taggers
