+++ python/dolma/__init__.py
@@ -12,7 +12,7 @@ from . import dolma as _dolma  # type: ignore   # noqa: E402
 from .core import TaggerRegistry  # noqa: E402
 from .core.errors import DolmaRustPipelineError  # noqa: E402
 from .core.taggers import BaseTagger  # noqa: E402
-from .taggers.language.en import *  # noqa: E402
+from .taggers.language import *  # noqa: E402
 
 __all__ = [
     "add_tagger",
