--- python/dolma/taggers/language/spa/code/__init__.py
@@ -1,15 +1,0 @@
-from .code_taggers import (
-    CodeCopyrightTagger,
-    CodeRedPajamaTaggers,
-    CodeSecretsTagger,
-    CodeStarCoderTaggers,
-    CodeStarCoderTaggers2,
-)
-
-__all__ = [
-    "CodeSecretsTagger",
-    "CodeCopyrightTagger",
-    "CodeRedPajamaTaggers",
-    "CodeStarCoderTaggers",
-    "CodeStarCoderTaggers2",
-]
